DatabaseLoader working against localdb attached from DB path file.
