+++ samples/csharp/getting-started/DatabaseLoader/DatabaseLoader.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28803.452
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LargeDatasetsInSqlServer", "LargeDatasetsInSqlServer\LargeDatasetsInSqlServer.csproj", "{DA4C03CF-0837-4C5C-88F9-8F6A365290E3}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DatabaseLoaderConsoleApp", "DatabaseLoaderConsoleApp\DatabaseLoaderConsoleApp.csproj", "{9058A1BB-87A6-4A9A-A3B2-F2C81CB9B9EA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{DA4C03CF-0837-4C5C-88F9-8F6A365290E3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DA4C03CF-0837-4C5C-88F9-8F6A365290E3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DA4C03CF-0837-4C5C-88F9-8F6A365290E3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DA4C03CF-0837-4C5C-88F9-8F6A365290E3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9058A1BB-87A6-4A9A-A3B2-F2C81CB9B9EA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9058A1BB-87A6-4A9A-A3B2-F2C81CB9B9EA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9058A1BB-87A6-4A9A-A3B2-F2C81CB9B9EA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9058A1BB-87A6-4A9A-A3B2-F2C81CB9B9EA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
