+++ samples/csharp/getting-started/DatabaseLoader/DatabaseLoaderConsoleApp/DataModels/ClickPrediction.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace DatabaseLoaderConsoleApp.DataModels
+{
+    public class ClickPrediction
+    {
+        public bool PredictedLabel;
+
+        public float Score;
+    }
+}
