+++ samples/csharp/getting-started/DatabaseLoader/DatabaseLoaderConsoleApp/DatabaseLoaderConsoleApp.csproj
@@ -1,0 +1,33 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <RestoreSources>
+      $(RestoreSources);
+    </RestoreSources>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.ML" Version="1.4.0-preview-28020-1" />
+    <PackageReference Include="Microsoft.ML.Experimental" Version="0.16.0-preview-28020-1" />
+    <PackageReference Include="Microsoft.ML.LightGBM" Version="1.4.0-preview-28020-1" />
+    <PackageReference Include="Microsoft.ML.FastTree" Version="1.4.0-preview-28020-1" />
+    <PackageReference Include="System.Data.SqlClient" Version="4.6.1" />
+    <PackageReference Include="TinyCsvParser" Version="2.0.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="SqlLocalDb\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="SqlLocalDb\Criteo-100k-rows.mdf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+    <None Update="SqlLocalDb\Criteo-100k-rows_log.ldf">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
+
+</Project>
