+++ samples/csharp/getting-started/DatabaseLoader/DatabaseLoaderConsoleApp/LargeDatasetsInSqlServer.csproj
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup />
+</Project>
