--- samples/csharp/getting-started/DatabaseLoader/LargeDatasetsInSqlServer/DataModels/ClickPrediction.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace LargeDatasetsInSqlServer.DataModels
-{
-    public class ClickPrediction
-    {
-        public bool PredictedLabel;
-
-        public float Score;
-    }
-}
