--- samples/csharp/getting-started/DatabaseLoader/LargeDatasetsInSqlServer/LargeDatasetsInSqlServer.csproj
@@ -1,20 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <RestoreSources>
-      $(RestoreSources);
-    </RestoreSources>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.ML" Version="1.4.0-preview-28020-1" />
-    <PackageReference Include="Microsoft.ML.Experimental" Version="0.16.0-preview-28020-1" />
-    <PackageReference Include="Microsoft.ML.LightGBM" Version="1.4.0-preview-28020-1" />
-    <PackageReference Include="Microsoft.ML.FastTree" Version="1.4.0-preview-28020-1" />
-    <PackageReference Include="System.Data.SqlClient" Version="4.6.1" />
-    <PackageReference Include="TinyCsvParser" Version="2.0.0" />
-  </ItemGroup>
-
-</Project>
